Refactoring  itens

+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/clientpackets/RequestDestroyItem.java
@@ -73,7 +73,7 @@ public final class RequestDestroyItem extends L2GameClientPacket
 			return;
 		}
 		
-		int count = _count;
+		long count = _count;
 		
 		if (activeChar.getPrivateStoreType() != 0)
 		{
