+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/clientpackets/RequestPackageSend.java
@@ -96,7 +96,7 @@ public final class RequestPackageSend extends L2GameClientPacket {
 
         // Freight price from config or normal price per item slot (30)
         int fee = _count * Config.ALT_GAME_FREIGHT_PRICE;
-        int currentAdena = player.getAdena();
+        long currentAdena = player.getAdena();
         int slots = 0;
 
         for (Item i : _items) {
@@ -177,14 +177,14 @@ public final class RequestPackageSend extends L2GameClientPacket {
             }
         }
 
-        // Send updated item list to the player
+        // Send updated item list to the reader
         if (playerIU != null) {
             player.sendPacket(playerIU);
         } else {
             player.sendPacket(new ItemList(player, false));
         }
 
-        // Update current load status on player
+        // Update current load status on reader
         StatusUpdate su = new StatusUpdate(player.getObjectId());
         su.addAttribute(StatusUpdate.CUR_LOAD, player.getCurrentLoad());
         player.sendPacket(su);
