+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/clientpackets/SendWareHouseDepositList.java
@@ -107,7 +107,7 @@ public final class SendWareHouseDepositList extends L2GameClientPacket
 		
 		// Freight price from config or normal price per item slot (30)
 		int fee = _count * 30;
-		int currentAdena = player.getAdena();
+		long currentAdena = player.getAdena();
 		int slots = 0;
 		
 		for (int i = 0; i < _count; i++)
@@ -205,7 +205,7 @@ public final class SendWareHouseDepositList extends L2GameClientPacket
 			}
 		}
 		
-		// Send updated item list to the player
+		// Send updated item list to the reader
 		if (playerIU != null)
 		{
 			player.sendPacket(playerIU);
@@ -215,7 +215,7 @@ public final class SendWareHouseDepositList extends L2GameClientPacket
 			player.sendPacket(new ItemList(player, false));
 		}
 		
-		// Update current load status on player
+		// Update current load status on reader
 		StatusUpdate su = new StatusUpdate(player.getObjectId());
 		su.addAttribute(StatusUpdate.CUR_LOAD, player.getCurrentLoad());
 		player.sendPacket(su);
